requirements.txt
streamlit==0.68.0
Pillow==8.0.1
